+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Init application services in parallel [#408](https://github.com/rokwire/app-flutter-plugin/issues/408).
 - SurveyPanel HeaderBar exposed for overriding [#4020](https://github.com/rokwire/illinois-app/issues/4020).
 ### Added
+- Added Group post scheduling [#4027](https://github.com/rokwire/illinois-app/issues/4027).
 - Added post type paramter to loadGroupPosts API [#441](https://github.com/rokwire/app-flutter-plugin/issues/441).
 - Created additional notifications for group post create/update/delete [#441](https://github.com/rokwire/app-flutter-plugin/issues/441).
 ### Added